add global message

+++ TwseAppFrontend/TwseApp/src/app/components/deleteDialog/delete-dialog/delete-dialog.component.ts
@@ -1,5 +1,6 @@
 import { Component, Inject } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { FormControl, Validators } from '@angular/forms';
+import {MatSnackBar} from '@angular/material/snack-bar';
 import { MatDialogRef } from '@angular/material/dialog';
 import { DeleteBrokerListResponse } from 'src/app/models/deleteBrokerListResponse';
 import { DeleteBrokerListService } from 'src/app/services/delete-company/delete-broker-list.service';
@@ -18,20 +19,31 @@ export interface DialogData {
 
 export class DeleteDialogComponent {
 
-  delete_code = new FormControl('');
-  constructor(public dialogRef: MatDialogRef<DeleteDialogComponent>,private deleteBrokerListService:DeleteBrokerListService) {}
+  delete_code = new FormControl('', [Validators.required]);
+  constructor(
+    public dialogRef: MatDialogRef<DeleteDialogComponent>
+    ,private deleteBrokerListService:DeleteBrokerListService
+    ,private _snackBar: MatSnackBar) {}
 
   onNoClick(): void {
     this.dialogRef.close();
   }
 
   TaskExecute():void {
-    if (!this.delete_code.value) return;
+    if (!this.delete_code.value || this.delete_code.value === "") return;
 
     this.deleteBrokerListService.deleteBrokerList(this.delete_code.value).subscribe({
       next:(response:DeleteBrokerListResponse) => {
-        console.log(response);
-        if (response.status) alert("Success");
+        if (response.status) {
+          this._snackBar.open('Delete Success', 'Info', {
+            duration: 3000
+          });
+        }
+        else{
+          this._snackBar.open('Delete Failed', 'Info', {
+            duration: 3000
+          });
+        }
       }
     })
